Feat: 불필요 코드 삭제 및 상태 호이스팅
